Schedule count on dashboard functionality
The schedule count comes from the amount of rows in the database table.
Note that as this is currently an in-memory database, the count is not
always accurate due to how and when the database is initialized.

Also note the temporary addition to App.js that corresponds to this
behavior.